refactoring complete
