+++ README.md
@@ -31,9 +31,9 @@ Load data: Gets data needed from database and loads a csv.-file of region codes 
 General method: parent of undated and dated methods.
 
 Undated_methods: Using sentiment analysed data from table 'data', calculates ratio of positive and negative sentiments grouped by region.
-Mainly consists of sorting as we are dealing with different datastructures(list of tuples, list of dicts, and pseudo-ordered list of dicts). It also maps all countries to their respective regions, which is needed to display them in the graph. This uploads duplicates, dunno how to fix.
+Mainly consists of sorting as we are dealing with different datastructures(list of tuples, list of dicts, and pseudo-ordered list of dicts). It also maps all countries to their respective regions, which is needed to display them in the graph. The ratio should be calculated using the previously calculated ratios. To do this, metadata is added before uploading it. Metada contains number of labels used to calculate mean and calculation date. A new mean can be calculated by multiplying the previous mean by previous labels used, adding the previous total to new total, adding previous labels to new labels, and calculating a new mean.
 
-Dated_methods: Does the same thing as undated_methods, but grouped by regions AND dates. The ratio should be calculated using the previously calculated ratios. To do this, metadata is added before uploading it. Metada contains number of labels used to calculate mean and calculation date. A new mean can be calculated by multiplying the previous mean by previous labels used, adding the previous total to new total, adding previous labels to new labels, and calculating a new mean.
+Dated_methods: Does the same thing as undated_methods, but grouped by regions AND dates. Does not calculate ratio using the old calculations in the same way, but does it freshly which is bad. This uploads duplicates, dunno how to fix.
 
 
 graph:
