+++ isitgoingtohell/data_management/db_management.py
@@ -1,13 +1,14 @@
 import psycopg2
-from isitgoingtohell.utils import load_json, dicts_to_tuples, tuple_to_dict, number_of_keys
-import os
-import pandas as pd
+from isitgoingtohell.utils import stringify_list, dicts_to_tuples, list_tuple_to_dict
 from psycopg2 import errors
 from psycopg2.errorcodes import UNIQUE_VIOLATION
-from datetime import date
 
 TABLENAME = 'data'
-class DB():
+UNDATED_TABLE = 'geography_undated'
+DATED_TABLE = 'geography'
+
+class Database():
+# Manage Connection
     def __init__(self):
         hostname = 'dpg-cdjur3un6mpngruf3uag-a.oregon-postgres.render.com'
         username = 'news_db_itmr_user'
@@ -20,49 +21,91 @@ class DB():
         # Create cursor, used to execute commands
         self.cur = self.connection.cursor()
 
-    def get_item_count(self, column, table, condition):
-        query = f"SELECT COUNT({column}) FROM {table} WHERE {condition}"
-        self.cur.execute(f"""{query} """)
+    def close_connection(self, message=True):
+        self.cur.close()
+        self.connection.close()
+        if message:
+            print("Connection closed. ")
+
+# Retrieve data
+    def get_data(self, columns_names='*', tablename=TABLENAME, condition='') -> list[tuple]:
+        # Input columns or condition as single string.
+        # Include WHERE for condition.
+        query = f" SELECT {columns_names} FROM {tablename} {condition}"
+        self.cur.execute(query)
+        return self.cur.fetchall()
+
+    def get_item_count(self, column, tablename=TABLENAME, condition=''):
+        # Input columns as single string, incluiding WHERE.
+        query = f"SELECT COUNT({column}) FROM {tablename} {condition}"
+        self.cur.execute(query)
         return self.cur.fetchone()
 
-    def upload_geography_undated(self, ud, columns):
-        query = f"INSERT INTO geography_undated ({columns}) values "
-        ratio = ud.calculate_ratio_total()
-        calculation_date = str(date.today().isoformat())
-        df = pd.DataFrame(ratio,columns=['region', 'score'])
-        ratio = df.to_dict('records')
-        new_order = []
-        for item in ratio:
-            new_item = {}
-            new_item['region'] = item['region']
-            new_item['score'] = item['score']
-            new_item['calculation_date'] = calculation_date
-            new_item['number_of_labels'] = self.get_item_count('score', 'data', f"region = '{item['region']}'")[0]
-            new_order.append(new_item)
+    def get_columns(self, include_id=False, tablename=TABLENAME) -> list:
+        query = f"SELECT * FROM {tablename} LIMIT 1"
+        self.cur.execute(query)
+        if include_id:
+            column_names = [col[0] for col in self.cur.description]
+            return column_names
+        else:
+            column_names = [col[0] for col in self.cur.description if col[0] != 'id']
+            return column_names
 
-        new_list = dicts_to_tuples(new_order)
+    def get_column_count(self, include_id=False, tablename=TABLENAME) -> int:
+        return len(self.get_columns(include_id, tablename))
 
-        for data in new_list:
-            self.cur.execute(f"""{query} """ + str(data))
-            print(f"{query}" + str(data))
+    def get_column_names(self, include_id=False, tablename=TABLENAME) -> list:
+        # Get names from desired substring of columns.
+        column_names = self.get_columns(include_id, tablename)
+        return column_names[:len(column_names)]
+        
+    def in_database(self, data: dict, tablename=TABLENAME) -> bool:
+        query = f" SELECT headline FROM {tablename} WHERE headline = '{data['headline']}' "
+        self.cur.execute(query)
+        if self.cur.fetchone():
+            return True
+        return False
+    
+    def get_unanalysed_data(self, tablename=TABLENAME):
+        condition = "WHERE label is Null"
+        column_names = self.get_column_names()
+        column_string = stringify_list(column_names)
+        data_tuples = self.get_data(column_string, tablename, condition)
+        return list_tuple_to_dict(data_tuples, column_names)
 
-        self.connection.commit()
+    def get_geography_data(self, dated=False, condition=""):
+        if dated:
+            tablename = DATED_TABLE
+        else:
+            tablename = UNDATED_TABLE
 
-    def upload_data_postgres_mogrify(self, mogrified_data, number_of_columns: int, tablename=TABLENAME):
-        # Insert the goodies to db. Cannot deal with duplicate data...
-        columns = self.get_col_names_not_id(tablename, number_of_columns)
-        query = f"insert into {tablename} ({columns}) values {mogrified_data}"
+        geo_data = self.get_data(tablename=tablename, condition=condition)
+        # Sorts data to fit for uploading
+        sorted_geo_data = self.sort_geo_data_for_database(geo_data, dated=dated)
+
+        return sorted_geo_data
+
+#Upload data
+    def insert_data(self, data: str, tablename=TABLENAME, columns_names='*', condition=''):
+        # Input columns or condition as single string.
+        # Include WHERE for condition.
+        query = f"INSERT INTO {tablename} ({columns_names}) VALUES {data} {condition}"
         self.cur.execute(query)
 
+    def insert_batch(self, mogrified_data: str, column_names: list, tablename=TABLENAME):
+        # WARNING: Cannot deal with duplicate data.
+        columns_string = ",".join(column_names)
+ 
+        self.insert_data(mogrified_data, tablename, columns_string)
+
         self.connection.commit()
 
-    def upload_data_postgres(self, data, number_of_columns: int, tablename=TABLENAME):
-        columns = self.get_col_names_not_id(tablename, number_of_columns)
-        ordered_columns = self.order_column_by_values(columns, data)
+    def insert_series(self, data: list[dict], number_of_columns: int, tablename=TABLENAME):
+        columns = self.get_column_names(tablename, number_of_columns)
+        ordered_columns = self.order_column_by_keys(columns, data)
         ordered_string = ",".join(ordered_columns)
-
         data = dicts_to_tuples(data)
-        # Insert the goodies to db.
+        
         counter = 0
         for nugget in data:
             headline = nugget[0].replace("'", "")    
@@ -76,69 +119,44 @@ class DB():
                 print(f"{e}")
                 self.connection.rollback()
         print(f"Uploaded {counter} item(s).")
+
         self.connection.commit()
 
-    def get_unanalysed_data(self, tablename=TABLENAME):
-        query=f"SELECT headline,date,region,label,score FROM {tablename} WHERE label is Null"
+    def update_one(self, data: list[dict], condition: str, tablename=TABLENAME):
+        # Updates one.
+        query = f"UPDATE {tablename} {condition}"
         self.cur.execute(query)
-        return self.cur.fetchall()
 
-    def upload_analysed_data(self, data, tablename=TABLENAME):
+    def upload_analysed_data(self, data: list[dict]):
         for item in data:
-            query = f"UPDATE {tablename} SET label = '{item['label']}', score = '{item['score']}' WHERE headline = '{item['headline']}'"
-            self.cur.execute(query)
+            condition = f"SET label = '{item['label']}', score = '{item['score']}' WHERE headline = '{item['headline']}'"
+            self.update_one(data, condition)
 
         self.connection.commit()
 
-    def order_column_by_values(self, columns, data) -> list:
-        # Implement a method that aligns the order of column and value 
-        # Warning: Keys and cols have to match.
-        dict_keys = data[0].keys()
-        ordered_keys = [key for key in dict_keys if key in columns]
-        return ordered_keys
+    def upload_geography_data(self, mogrified_data: str, dated=False):
+        if dated:
+            tablename = DATED_TABLE
+        else:
+            tablename = UNDATED_TABLE
+        column_names = self.get_column_names(tablename=tablename)
+        self.insert_batch(mogrified_data, column_names=column_names, tablename=tablename)
+        self.connection.commit()
 
-    def mogrify_data(self, data, columns):
+# Modify data
+    def mogrify_data(self, data: list[dict], number_of_columns: int) -> str:
         # Turn list of dicts data into a single query string
-        values = []
+        ## Mogrify method needs list of tuples. We shall provide.
+        values = dicts_to_tuples(data)
 
-        # Mogrify method needs list of tuples. We shall provide.
-        for row in data:
-            values.append(tuple(row.values()))
-        x = 0
-        format = []
-        while x < columns:
-            format.append('%s')
-            x+=1
-        format = ','.join(format)    
+        format = self.string_format(number_of_columns)    
 
         value_string = ','.join(self.cur.mogrify(f"({format})", row).decode('utf-8') for row in values)
 
         return value_string
 
-    def get_col_names_not_id(self, tablename, number_of_columns: int, outbound=False) -> str:
-        # Get names from col except for id
-        self.cur.execute(f"""select * from {tablename}""")
-
-        # Gets all col names except for id, stopping at columns
-        column_names_list = [col[0] for col in self.cur.description if col[0] != 'id']
-        column_names_string = ",".join(column_names_list[:number_of_columns])
-
-        if outbound:
-            self.close_connection(message=False)
-
-        return column_names_string
-
-    def get_col_data(self, tablename, column_name, outbound=False) -> list:
-        # Create list of all headlines in database. Probably doesn't scale great, but is quick.
-        self.cur.execute(f""" select {column_name} from {tablename} """)
-        col_names = [i[0] for i in self.cur.fetchall()]
-
-        if outbound:
-            self.close_connection(message=False)
-
-        return col_names
-
-    def remove_duplicates_batch(self, data, tablename=TABLENAME):
+    def remove_duplicates_batch(self, data: list[dict], tablename=TABLENAME)->list:
+        # Probably specify name on these
         headlines = [("headline = " + "'" +data['headline']+ "'") for data in data]
         condition = " OR ".join(headlines)
         query = f" SELECT headline FROM {tablename} WHERE {condition} "
@@ -146,11 +164,20 @@ class DB():
         duplicates = [headlines[0] for headlines in self.cur.fetchall()]
         uniques = [item for item in data if item['headline'] not in duplicates]
         return uniques
-
-    def remove_duplicates_one_by_one(self, data, tablename=TABLENAME):
+        
+    def remove_duplicate(self, data: list[dict], tablename=TABLENAME):
         #Goes through each item in data and checks if it is already in database. Safe, but slow.
         items_not_in_database = [item for item in data if not self.in_database(item)]
         return items_not_in_database
+    
+    def string_format(self, number_of_columns:int) -> str:
+        # Takes number of columns and strings together '%s'.
+        x = 0
+        format = []
+        while x < number_of_columns:
+            format.append('%s')
+            x+=1
+        return ','.join(format) 
 
     def in_database(self, data: dict, tablename=TABLENAME) -> bool:
         # Check if item exists in database.
@@ -159,77 +186,10 @@ class DB():
         if self.cur.fetchone():
             return True
         return False
-        
-    def get_geography_data(self):
-        raw_data = self.get_all_data("geography")
-        data = []
-        for x in raw_data:
-            d = {}
-            d['country_code'] = x[0]
-            d['score'] = x[1]
-            d['date'] = x[2]
-            d['region'] = x[3]
-            data.append(d)
-        return data
 
-    def get_geography_data_undated(self):
-        raw_data = self.get_all_data("geography_undated")
-        data = []
-        for x in raw_data:
-            d = {}
-            d['country_code'] = x[0]
-            d['region'] = x[1]
-            d['score'] = x[2]
-            d['calculation_date'] = x[3]
-            d['number_of_labels'] = x[4]
-            data.append(d)
-        return data
-
-    def get_all_data(self, tablename:str):
-        self.cur.execute(f""" select * from {tablename} """)
-        return self.cur.fetchall()
-
-    def get_analysed_data(self, tablename:str):
-        query=f"SELECT headline,date,region,label,score FROM {tablename} WHERE label is not Null"
-        self.cur.execute(query)
-        return self.cur.fetchall()
-
-    def upload_geography_undated(self, columns):
-        query = f"INSERT INTO geography_undated ({columns}) values "
-        ratio = self.calculate_ratio_total()
-        calculation_date = str(date.today().isoformat())
-        df = pd.DataFrame(ratio,columns=['region', 'score'])
-        ratio = df.to_dict('records')
-        new_order = []
-        for item in ratio:
-            new_item = {}
-            new_item['region'] = item['region']
-            new_item['score'] = item['score']
-            new_item['calculation_date'] = calculation_date
-            new_item['number_of_labels'] = self.get_item_count('score', 'data', f"region = '{item['region']}'")[0]
-            new_order.append(new_item)
-
-        new_list = dicts_to_tuples(new_order)
-
-        for data in new_list:
-            self.cur.execute(f"""{query} """ + str(data))
-            print(f"{query}" + str(data))
-
-        self.connection.commit()
-
-    def upload_populated_regions(self, columns, region_values):
-        query = f"INSERT INTO geography ({columns}) values "
-
-        new_list = dicts_to_tuples(region_values)
-
-        for data in new_list:
-            self.cur.execute(f"""{query} """ + str(data))
-            print(f"{query}" + str(data))
-
-        self.connection.commit()
-
-    def close_connection(self, message=True):
-        self.cur.close()
-        self.connection.close()
-        if message:
-            print("Connection closed. ")
+    def order_column_by_keys(self, columns: list, data: dict) -> list:
+        # Changes order if colums to match the keys of a dictionary. 
+        # WARNING: Keys and columns have to match namewise.
+        dict_keys = data.keys()
+        ordered_keys = [key for key in dict_keys if key in columns]
+        return ordered_keys
