+++ isitgoingtohell/sentiment_analysis/sentiment_analysis.py
@@ -1,4 +1,4 @@
-from isitgoingtohell.utils import load_toml, load_json
+from isitgoingtohell.utils import load_toml
 from isitgoingtohell.data_management.db_management import DB
 from transformers import pipeline
 import numpy as np
